Refactor organization ports to dedicated adapters
