+++ docs/architecture.md
@@ -16,7 +16,7 @@ review_cadence: quarterly
 
 ## Feature Catalog and Flows
 - **Metadata** (`src/omym/features/metadata`): extracts and enriches tags via `MusicProcessor`, coordinates optional MusicBrainz lookups, emits processing events consumed by organisation flows, and surfaces database/cache ports in `usecases/ports.py` so adapters remain swappable. The use case now composes dedicated helper packages (`usecases/assets/` for lyrics and artwork flows, `usecases/processing/` for orchestration, cleanup, and shared result types) to keep each module under 300 LOC while preserving the existing public API.
-- **Organization** (`src/omym/features/organization`): groups tracks into album/disc hierarchies, evaluates user-defined filters, composes metadata and path services to derive final layouts, and exposes album/filter repository ports in `usecases/ports.py` for adapter injection.
+- **Organization** (`src/omym/features/organization`): groups tracks into album/disc hierarchies, evaluates user-defined filters, composes metadata and path services to derive final layouts, and exposes album/filter repository ports in `usecases/ports.py` with SQLite adapters now living under `adapters/`.
 - **Path** (`src/omym/features/path`): generates filesystem-safe directory and filename structures, sanitises inputs, exposes helpers that organisation and restoration use to resolve canonical targets, and now publishes filter access ports in `usecases/ports.py`.
 - **Restoration** (`src/omym/features/restoration`): reconstructs original locations from persisted processing plans, applies collision policies, and coordinates filesystem/database adapters for rollback.
 - Features collaborate exclusively through ports defined in their `usecases` modules. For example, the organisation use cases consume `MetadataExtractor` and path generation ports rather than importing foreign domains directly.
