+++ src/omym/features/organization/adapters/__init__.py
@@ -1,0 +1,9 @@
+"""
+Summary: Adapter exports for the organization feature.
+Why: Provide ready-to-use infrastructure bridges for album and filter ports.
+"""
+
+from .album_repository_adapter import AlbumDaoAdapter
+from .filter_registry_adapter import FilterDaoAdapter
+
+__all__ = ["AlbumDaoAdapter", "FilterDaoAdapter"]
