+++ src/omym/features/organization/usecases/manage_albums.py
@@ -1,15 +1,10 @@
-"""Album management use cases.
-
-Where: features/organization/usecases.
-What: Group processed files into albums using persistence ports.
-Why: Keep SQLite-specific behaviour inside adapters for better testability.
+"""
+Summary: Album management use cases depending solely on repository ports.
+Why: Maintain hexagonal separation by moving DAO construction into adapters.
 """
 
 from dataclasses import dataclass
-from sqlite3 import Connection
 from typing import final
-
-from omym.platform.db.daos.albums_dao import AlbumDAO
 
 from .ports import AlbumRecord, AlbumRepositoryPort
 
@@ -29,24 +24,13 @@ class AlbumManager:
 
     _albums: AlbumRepositoryPort
 
-    def __init__(
-        self,
-        conn: Connection | None = None,
-        *,
-        album_port: AlbumRepositoryPort | None = None,
-    ) -> None:
+    def __init__(self, album_port: AlbumRepositoryPort) -> None:
         """Initialize album manager.
 
         Args:
-            conn: Legacy SQLite connection used when ``album_port`` is omitted.
             album_port: Port that manages album persistence.
         """
-        if album_port is not None:
-            self._albums = album_port
-        elif conn is not None:
-            self._albums = _AlbumDaoAdapter(conn)
-        else:
-            raise ValueError("AlbumManager requires album_port or conn")
+        self._albums = album_port
 
     def process_files(self, files: dict[str, dict[str, str | None]]) -> tuple[list[AlbumGroup], list[str]]:
         """Process files and group them into albums.
@@ -270,44 +254,3 @@ class AlbumManager:
         return earliest_year
 
 
-@final
-class _AlbumDaoAdapter(AlbumRepositoryPort):
-    """Adapter projecting AlbumDAO behaviour through the album repository port."""
-
-    def __init__(self, conn: Connection) -> None:
-        self._dao = AlbumDAO(conn)
-
-    def get_album(self, album_name: str, album_artist: str) -> AlbumRecord | None:
-        record = self._dao.get_album(album_name, album_artist)
-        if record is None:
-            return None
-        return AlbumRecord(
-            id=record.id,
-            album_name=record.album_name,
-            album_artist=record.album_artist,
-            year=record.year,
-            total_tracks=record.total_tracks,
-            total_discs=record.total_discs,
-        )
-
-    def insert_album(
-        self,
-        album_name: str,
-        album_artist: str,
-        year: int | None = None,
-        total_tracks: int | None = None,
-        total_discs: int | None = None,
-    ) -> int | None:
-        return self._dao.insert_album(album_name, album_artist, year, total_tracks, total_discs)
-
-    def insert_track_position(
-        self,
-        album_id: int,
-        disc_number: int,
-        track_number: int,
-        file_hash: str,
-    ) -> bool:
-        return self._dao.insert_track_position(album_id, disc_number, track_number, file_hash)
-
-    def check_track_continuity(self, album_id: int) -> tuple[bool, list[str]]:
-        return self._dao.check_track_continuity(album_id)
