+++ src/omym/features/organization/usecases/manage_filters.py
@@ -1,17 +1,12 @@
-"""Filtering engine use cases.
-
-Where: features/organization/usecases.
-What: Manage hierarchy registration and per-file filter values via injected ports.
-Why: Keep database specifics outside the use case for easier testing and adapter swaps.
+"""
+Summary: Filtering engine use cases relying on injected ports only.
+Why: Enforce hexagonal boundaries by removing direct DAO construction in the use case.
 """
 
-from sqlite3 import Connection
 from typing import final
 
-from omym.platform.db.daos.filter_dao import FilterDAO
-
 from ..domain.path_format import parse_path_format
-from .ports import FilterHierarchyRecord, FilterRegistryPort
+from .ports import FilterRegistryPort
 
 @final
 class HierarchicalFilter:
@@ -19,24 +14,13 @@ class HierarchicalFilter:
 
     _filters: FilterRegistryPort
 
-    def __init__(
-        self,
-        conn: Connection | None = None,
-        *,
-        filter_port: FilterRegistryPort | None = None,
-    ) -> None:
-        """Initialize filter engine.
+    def __init__(self, filter_port: FilterRegistryPort) -> None:
+        """Initialize filter engine with a persistence port.
 
         Args:
-            conn: Legacy SQLite connection used when ``filter_port`` is omitted.
             filter_port: Port implementation providing hierarchy persistence.
         """
-        if filter_port is not None:
-            self._filters = filter_port
-        elif conn is not None:
-            self._filters = _FilterDaoAdapter(conn)
-        else:
-            raise ValueError("HierarchicalFilter requires filter_port or conn")
+        self._filters = filter_port
 
     def register_hierarchies(self, path_format: str) -> list[str]:
         """Register filter hierarchies from path format.
@@ -82,21 +66,3 @@ class HierarchicalFilter:
         return warnings
 
 
-@final
-class _FilterDaoAdapter(FilterRegistryPort):
-    """Adapter projecting FilterDAO into the filter registry port."""
-
-    def __init__(self, conn: Connection) -> None:
-        self._dao = FilterDAO(conn)
-
-    def insert_hierarchy(self, name: str, priority: int) -> int | None:
-        return self._dao.insert_hierarchy(name, priority)
-
-    def get_hierarchies(self) -> list[FilterHierarchyRecord]:
-        return [
-            FilterHierarchyRecord(id=item.id, name=item.name, priority=item.priority)
-            for item in self._dao.get_hierarchies()
-        ]
-
-    def insert_value(self, hierarchy_id: int, file_hash: str, value: str) -> bool:
-        return self._dao.insert_value(hierarchy_id, file_hash, value)
